Styled Components III (Repository)
-Index from Repository EDITED
  -Included a Header section that contains an <img> with a logo and
a <svg> FiChevronLeft icon to return to the previous page.
  -Included a Repository Summary section that contains the selected
repository's <img> and description (<span). It also contains a <ul>
with the number of star, forks and open issues as list items.
  -Included an Issues section that contains a link (<a>) to the issue's details.
<a> has a <strong> that contains the issue's title, a <p> that contains a brief
description about the issue, and a <svg> FiChevronRight icon.

Styles from Repository EDITED
  -Included styling for the individual parts of the Repository
  -Included a hover transition for <FiChevronLeft> that takes 0.2 seconds
to darken its color shade.
  -Included a hover transition for <a> that takes 0.2 seconds to move it 10px
in the X Axis and it also darkens the <FiChevronRight> component's color shade.